refactor: car 패키지로 이동 & final로 변수 선언 변경
