+++ src/main/java/racing/domain/car/CarHistory.java
@@ -1,0 +1,20 @@
+package racing.domain.car;
+
+public class CarHistory {
+
+    private final Name name;
+    private final Position position;
+
+    public CarHistory(Name name, Position position) {
+        this.name = new Name(name);
+        this.position = new Position(position);
+    }
+
+    public Name getName() {
+        return name;
+    }
+
+    public Position getPosition() {
+        return position;
+    }
+}
